Major cleanup - renamed ViTImageProcessor to ViTFeatureExtractor
