+++ src/transformers/models/vit/__init__.py
@@ -17,13 +17,15 @@
 # limitations under the License.
 from typing import TYPE_CHECKING
 
-from ...file_utils import _BaseLazyModule, is_tokenizers_available, is_torch_available
+from ...file_utils import _BaseLazyModule, is_torch_available, is_torchvision_available
 
 
 _import_structure = {
     "configuration_vit": ["VIT_PRETRAINED_CONFIG_ARCHIVE_MAP", "ViTConfig"],
-    "image_processor_vit": ["ViTImageProcessor"],
 }
+
+if is_torchvision_available():
+    _import_structure["feature_extraction_vit"] = ["ViTFeatureExtractor"]
 
 if is_torch_available():
     _import_structure["modeling_vit"] = [
@@ -37,8 +39,10 @@ if is_torch_available():
 
 if TYPE_CHECKING:
     from .configuration_vit import VIT_PRETRAINED_CONFIG_ARCHIVE_MAP, ViTConfig
-    from .image_processor_vit import ViTImageProcessor
 
+    if is_torchvision_available():
+        from .feature_extraction_vit import ViTFeatureExtractor
+    
     if is_torch_available():
         from .modeling_vit import (
             VIT_PRETRAINED_MODEL_ARCHIVE_LIST,
