+++ tests/test_modeling_vit.py
@@ -17,6 +17,7 @@
 
 import unittest
 
+import torchvision
 import torchvision.transforms as T
 from PIL import Image
